Modify implementation of database servlets.

+++ qualitypatternmodel/src/qualitypatternmodel/newservlets/DatabaseDeleteServlet.java
@@ -1,6 +1,8 @@
 package qualitypatternmodel.newservlets;
 
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 
 import jakarta.servlet.http.HttpServlet;
@@ -56,9 +58,23 @@ public class DatabaseDeleteServlet extends HttpServlet {
 		if (!ServletUtilities.TECHS.contains(technology))
 			throw new InvalidServletCallException("invalid technology");
 		
-		// TODO
-		// check if any constraint uses that database
-		// -> Failed 'Database in Use by x constrains: [..]
+		if (ServletUtilities.loadDatabase(technology, dbname) == null)
+			throw new FailedServletCallException("Database does not exist.");
+		
+//		TODO: database.getName()
+//		try {
+//			List<CompletePattern> patterns = ServletUtilities.getAllSemiConcretePattern(technology);
+			List<String> patternnames = new ArrayList<String>();
+			
+//			for (CompletePattern pattern: patterns) {
+//				if (pattern.getDatabase().getName().equals(dbname)) {
+//					patternnames.add(pattern.getName());
+//				}
+//			}
+			if (!patternnames.isEmpty()) {
+				throw new FailedServletCallException("Database is in use within the following constraints: " + patternnames);
+			}
+//		} catch (IOException e) {}
 		
 		ServletUtilities.deleteDatabase(technology, dbname);
